zmq-uploader/requirements.txt
pyzmq
PyQt5
grbl2image
pillow